requirements.txt
pyinstaller==5.0.1
python-decouple==3.8
requests==2.25.1